packages: SystemUI: improve the VPN footer layout
Not only is the background color of the VPN footer layout a different color
but the text is not centered. The padding was just off.

- Also cloned the styles for this layout and made the text bold. Google clearly
wants the user to know that they're connected to a VPN by adding a statusbar
icon and giving this its own footer. This just adds to that by making it stick
out slightly.

Signed-off-by: iamtoxyc <vtechnical1804@gmail.com>